% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psidyonsei.R
\name{psid_read2}
\alias{psid_read2}
\title{Alternative psid_read() function}
\usage{
psid_read2(indir, str_df, idvars = NA, type, filename = NA)
}
\value{
well-organized string of vectors to your clipboard
}
\description{
If 'psid_read' function from the 'psid_read' package causes an error, try this instead.(check psid_read help file)
}
