% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psidyonsei.R
\name{var_search}
\alias{var_search}
\title{Help search PSID variables}
\usage{
var_search(varlist)
}
\arguments{
\item{varlist}{A Vector of Variables copied from PSID codebook}
}
\value{
well-organized string of vectors to your clipboard
}
\description{
After you created variable list using 'load_var', use 'var_search' to help search in the PSID website.
}
\examples{
var_search(c("state || [68]V93 [69]V537 [70]V1103 [71]V1803 [72]V2403 [73]V3003 [74]V3403 [75]V3803 [76]V4303 [77]V5203 [78]V5703 [79]V6303 [80]V6903 [81]V7503 [82]V8203 [83]V8803 [84]V10003 [85]V11103 [86]V12503 [87]V13703 [88]V14803 [89]V16303 [90]V17703 [91]V19003 [92]V20303 [93]V21603 [94]ER4156 [95]ER6996 [96]ER9247 [97]ER12221 [99]ER13004 [01]ER17004 [03]ER21003 [05]ER25003 [07]ER36003 [09]ER42003 [11]ER47303 [13]ER53003 [15]ER60003 [17]ER66003 [19]ER72003 [21]ER78003"))
}
